+++ changelog.md
@@ -1,3 +1,10 @@
+1.4.2:
+    Cata:
+    - Fixed darkmode not applying to focus tot
+
+    Retail:
+    - Shortened ToT name so it does not go outside of the frame texture.
+
 Cataclysm 1.4.1e:
 - Fixed logic with target/focus auras messing up with eachother after port from retail to cata.
 - Fixed Name Bg setting on Player to only be the actual size of the name bg so color behind hp/mana doesnt get changed